% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vector_constraint.R
\name{build_vector_constraint}
\alias{build_vector_constraint}
\title{build_vector_constraint
internal function that convert a symbolic equation of constraint into a matrix storing the coefficients of the constraint}
\usage{
build_vector_constraint(eq_constraint)
}
\arguments{
\item{eq_constraint}{a symbolic equation corresponding to a constraint}
}
\value{
a sparse matrix storing the coefficient of the constraint (first column stands for -intercept)
}
\description{
build_vector_constraint
internal function that convert a symbolic equation of constraint into a matrix storing the coefficients of the constraint
}
