% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCaN.R
\name{buildCaN}
\alias{buildCaN}
\title{buildCaN}
\usage{
buildCaN(x, generic = FALSE)
}
\arguments{
\item{x}{either the path to a RCaN input file or a named list with
at least 4 elements (components_param, fluxes_def, series, constraints). It
can also includes an element dynamics if this is not a trophic model, and
optionally aliases}

\item{generic}{tells whether the model is a standard RCaN trophic model
(FALSE, default) or a generic model (TRUE)}
}
\value{
a CaNmod object with following elements
\itemize{
 \item{"components_param"}{the table of components description}
 \item{"species"}{the name of the species}
 \item{"fluxes_def}{the table of fluxes definition}
 \item{"ntstep"}{the number of time steps}
 \item{"aliases"}{the aliases specified by the user}
 \item{"dynamics"}{the dynamics equation for a generic model}
 \item{"data_series_name"}{the names of the data series}a
 \item{"constraints"}{the table of constraints description}
 \item{"H"}{the H matrix from (I-H).B+N.F}
 \item{"N"}{the N matrix from (I-H).B+N.F}
 \item{"A"}{matrix of active constraints A.x<=b}
 \item{"AAll"}{matrix of all constraints A.x<=b}
 \item{"b"}{vector of active constraints A.x<=b}
 \item{"bAll"}{vector of all constraints A.x<=b}
 \item{"C"}{matrix of active constraints C.x=v}
 \item{"CAll"}{matrix of all constraints C.x=v}
 \item{"v"}{vector of active constraints C.x=v}
 \item{"vAll"}{vector of all constraints C.x=v}
 \item{"L"}{matrix of B=L.F, since B0 is a parameter, there is no M}
 \item{"symbolic_enviro"}{an environment storing all symbolic objects
 required for the computation}
}
}
\description{
builds a CaNmodobject directly either by reading an RCaN input file or by
using a list of dataframes instead of and returns the complete model
description, including all the underlying equations
}
\examples{
require(readxl)
file <- system.file("extdata",
 "CaN_template_mini.xlsx", package = "RCaN")

 #for this example, we take the table as provided in the template file
 #but you can provide your own table form other sources

 #read compartments
  components_param <-
read_excel(file, sheet = "Components & input parameter")

#read Fluxes
fluxes_def <- read_excel(file, sheet = "Fluxes")

#read Times series
series <- read_excel(file, sheet = "Input time-series")

#read constraints
constraints <- read_excel(file, sheet = "Constraints")

myCaNmod <- buildCaN(list(components_param = components_param,
                               fluxes_def = fluxes_def,
                               series = series,
                               constraints = constraints))

#example with a template under the generic format
file <- system.file("extdata",
 "CaN_template_mini_generic.xlsx", package = "RCaN")
myCaNmod_generic <- buildCaN(file, generic = TRUE)

}
