% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildVectorConstraint.R
\name{buildVectorConstraint}
\alias{buildVectorConstraint}
\title{buildVectorConstraint
internal function that convert a symbolic equation of constraint into a
matrix storing the coefficients of the constraint}
\usage{
buildVectorConstraint(eq_constraint, symbolic_enviro)
}
\arguments{
\item{eq_constraint}{a symbolic equation corresponding to a constraint}

\item{symbolic_enviro}{the symbolic environment}
}
\value{
a sparse matrix storing the coefficient of the constraint (first
column stands for -intercept)
}
\description{
buildVectorConstraint
internal function that convert a symbolic equation of constraint into a
matrix storing the coefficients of the constraint
}
