% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyCentre.R
\name{chebyCentre}
\alias{chebyCentre}
\alias{chebyCenter}
\title{chebyCentre
Computes the centroid of a convex polytope. chebyCentre and chebyCenter are
synonyms.}
\usage{
chebyCentre(A, b)

chebyCenter(A, b)
}
\arguments{
\item{A}{a matrix}

\item{b}{a vector of length equals to nrow(A)}
}
\value{
a vector corresponding to the centroid of the polytope
}
\description{
chebyCentre
Computes the centroid of a convex polytope. chebyCentre and chebyCenter are
synonyms.
}
\section{Details}{

This code is a translation of the matlab code that can be found there
(https://bit.ly/3b6Avav)
It computes the centroid of the complex polytope defined by
\eqn{A \cdot x \leqslant   b}
}

\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- chebyCentre(A,b)
}
