% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPolytopeStatusCaNmod.R
\name{checkPolytopeStatusCaNmod}
\alias{checkPolytopeStatusCaNmod}
\title{checkPolytopeStatusCaNmod
check whether the polytope is well defined or not}
\usage{
checkPolytopeStatusCaNmod(myCaNmod)
}
\arguments{
\item{myCaNmod}{a CaNmod oject}
}
\value{
print a message to tell if the polygon is ok or not
}
\description{
checkPolytopeStatusCaNmod
check whether the polytope is well defined or not
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
checkPolytopeStatusCaNmod(myCaNmod)

#we artificially add incompatible constraints (negative flow)
myCaNmod$A <- rbind(myCaNmod$A,c(1,rep(0,ncol(myCaNmod$A)-1)))
rownames(myCaNmod$A)[nrow(myCaNmod$A)]<-"neg_flow"
myCaNmod$b <- c(myCaNmod$b,-1)
checkPolytopeStatusCaNmod(myCaNmod)
}
