% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDynamics.R
\name{createDynamics}
\alias{createDynamics}
\title{createDynamics}
\usage{
createDynamics(dynamics_equation, components, fluxes)
}
\arguments{
\item{dynamics_equation}{a string that specifies the dynamics equation}

\item{components}{the components data frame}

\item{fluxes}{the fluxes data frame}
}
\value{
a list with H and N matrices
}
\description{
builds H and N matrix that relate biomass at time step t+1 to fluxes and
biomasses at time t in the case of non standard RCaN model. This function
is not called by user but by buildCaN.
}
