% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineLPMod.R
\name{defineLPMod}
\alias{defineLPMod}
\title{defineLPMod
Specify an lp_model (mainly for internal use)}
\usage{
defineLPMod(A, b, C = NULL, v = NULL, presolve = c("rows", "lindep",
  "cols"), lower = NULL)
}
\arguments{
\item{A}{a matrix of inequality constraints A x <=b}

\item{b}{a vector A x <=b}

\item{C}{a matrix of equality constraints C x = v, should be null in the
absence of such constraints}

\item{v}{a vector of equality constraints C x = v, should be null in the
absence of such constraints}

\item{presolve}{argument send to \code{\link[lpSolveAPI]{lp.control}}}

\item{lower}{minimal bounds for paramaters, by default set to zero}
}
\value{
a vector corresponding to the centroid of the polytope
}
\description{
defineLPMod
Specify an lp_model (mainly for internal use)
}
