% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateInertia.R
\name{estimateInertia}
\alias{estimateInertia}
\title{estimateInertia
Estimate inertia based on allometric relationships}
\usage{
estimateInertia(
  weight,
  method,
  taxon,
  temperature = NULL,
  d = 1,
  fractional = 0.1
)
}
\arguments{
\item{weight}{wet weight in gram}

\item{method}{one of "savage" or "yodzis"}

\item{taxon}{one of ("endotherm", #' "vertebrate ecotherm", "invertebrate")
for yodzis and one of ("unicellular eukaryote","multicellular eukaryote",
"vertebrate") for savaage}

\item{temperature}{temperature in °C (if savage)}

\item{d}{fraction of days in which a species grow (1 for 365 days / year).
used in savage}

\item{fractional}{Fractional properties (unitless), generally in the order
of 0.1 (default value)}
}
\value{
satiation as explained in supplementary material in
\insertCite{planque2022;textual}{RCaN}
}
\description{
estimateInertia
Estimate inertia based on allometric relationships
}
\details{
Following \insertCite{planque2022;textual}{RCaN}, this function
estimates satiation based on an allometric relationship formulated by
\insertCite{yodzis1992;textual}{RCaN} or
\insertCite{savage2004;textual}{RCaN}
}
\examples{
estimateInertia(1e-4, "yodzis", taxon = "invertebrate", fractional = 0.1)
}
\references{
\insertAllCited{}
}
