% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmyCaNmod.R
\name{fitmyCaNmod}
\alias{fitmyCaNmod}
\title{fitmyCaNmod
fit the CaNmod model}
\usage{
fitmyCaNmod(myCaNmod, N)
}
\arguments{
\item{myCaNmod}{a CaNmod object with following elements}

\item{N}{the number of samples required}
}
\value{
a list with F and B the matrices of biomass and flows
}
\description{
fitmyCaNmod
fit the CaNmod model
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx", package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)

}
