% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllBoundsParam.R
\name{getAllBoundsParam}
\alias{getAllBoundsParam}
\title{getAllBoundsParam
Computes the possible bounds for all parameters of a polytope defined by
A.x<=b and C.x=v}
\usage{
getAllBoundsParam(A, b, C = NULL, v = NULL)
}
\arguments{
\item{A}{the matrix of inequality A.x<=b}

\item{b}{the vector A.x<=b}

\item{C}{the matrix of equality C.x=v (default NULL for no equality)}

\item{v}{the vector of equality C.x=v (default NULL for no equality}
}
\value{
a datrame with first column corresponding to colnames(A), and
corresponding lower bounds (column 2) and upper bounds (column 3)
}
\description{
getAllBoundsParam
Computes the possible bounds for all parameters of a polytope defined by
A.x<=b and C.x=v
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- getAllBoundsParam(A,b)
}
