% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllBoundParamCaNmod.R
\name{getAllBoundsParamCaNmod}
\alias{getAllBoundsParamCaNmod}
\title{getAllBoundsParamCaNmod
compute bounds for all paramaters}
\usage{
getAllBoundsParamCaNmod(myCaNmod)
}
\arguments{
\item{myCaNmod}{a CaNmod oject}
}
\value{
a data frame with first column as names of params and 2 columns for
bounds
}
\description{
getAllBoundsParamCaNmod
compute bounds for all paramaters
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
getAllBoundsParamCaNmod(myCaNmod)
}
