% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundParam.R
\name{getBoundParam}
\alias{getBoundParam}
\title{getBoundParam
Computes the possible bounds for a parameter p of a polytope defined by+A.x<=b and C.x=v}
\usage{
getBoundParam(A, b, p, C = NULL, v = NULL)
}
\arguments{
\item{A}{the matrix of inequality A.x<=b}

\item{b}{the vector A.x<=b}

\item{p}{the index of the parameter for which the bounds should be computed}

\item{C}{the matrix of equality C.x=v (default NULL for no equality)}

\item{v}{the vector of equality C.x=v (default NULL for no equality}
}
\value{
a vector with lower bounds and upper bounds
}
\description{
getBoundParam
Computes the possible bounds for a parameter p of a polytope defined by+A.x<=b and C.x=v
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- getBoundParam(A,b,1)

}
