% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundParamCaNmod.R
\name{getBoundParamCaNmod}
\alias{getBoundParamCaNmod}
\title{getBoundParamCaNmod
compute bounds for the paramater p}
\usage{
getBoundParamCaNmod(myCaNmod, p)
}
\arguments{
\item{myCaNmod}{a CaNmod oject}

\item{p}{the paramter to be analysed}
}
\value{
a vector with lower bound and upper bound
}
\description{
getBoundParamCaNmod
compute bounds for the paramater p
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
# with the index
getBoundParamCaNmod(myCaNmod,1)
# with the name
getBoundParamCaNmod(myCaNmod,"F01[1988]")
}
