% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamMinMax.R
\name{getParamMinMax}
\alias{getParamMinMax}
\title{getParamMinMax
Computes the minum value bounds for a parameter p of a polytope defined
by+A.x<=b and C.x=v or by the polytope of the CaNmod object}
\usage{
getParamMinMax(x, p, presolved, maximum)
}
\arguments{
\item{x}{either a CaNmod object or a named list with at least a matrix A and
a vector b (A.x<=b) and optionally a matrix C and a vector v (C.x=v)}

\item{p}{the index of the parameter for which the bounds should be computed}

\item{presolved}{an object returned by \link{presolveLPMod}}

\item{maximum}{should ob be maximized}
}
\value{
a vector with lower bounds and upper bounds
}
\description{
getParamMinMax
Computes the minum value bounds for a parameter p of a polytope defined
by+A.x<=b and C.x=v or by the polytope of the CaNmod object
}
