% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBottleneck.R
\name{ggBottleneck}
\alias{ggBottleneck}
\title{ggBottleneck
display whether species were closed to satiation or if their predators
were closed to inertia}
\usage{
ggBottleneck(mysampleCaNmod, species = NULL, years = NULL)
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{species}{the species to plot (if null, default, all species)}

\item{years}{years to be plotted (default all)}
}
\value{
plots in a grid.arrange
}
\description{
ggBottleneck
display whether species were closed to satiation or if their predators
were closed to inertia
}
\details{
the idea of this plot is to show whether a species has enough food and is or
limiting in its predator diet. As such, the x-axis displays how far the
species is from satiation (1 satiation, 0 no food) and the y-axis displays
how far the sum of trophics flow towards its predator is from the flows that
would have occured if all predators had achieved satiation (1 = all predators
are at satiation). Each point stands for a time step in one iteration of
MCMC samples.
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaN(myCaNmod, 100)
ggBottleneck(res)
}
