% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGrowth.R
\name{ggGrowth}
\alias{ggGrowth}
\title{ggGrowth
plots biomass growth as a function of biomass
provides a distribution over iterations and years}
\usage{
ggGrowth(myFitCaNmod, species = NULL)
}
\arguments{
\item{myFitCaNmod}{result sent by \link{fitmyCaNmod}}

\item{species}{the name (or a vector of name) of the species of interest
by default, all species}
}
\value{
a ggplot
}
\description{
ggGrowth
plots biomass growth as a function of biomass
provides a distribution over iterations and years
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)
ggGrowth(res)

}
