% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPairsBiomass.R
\name{ggPairsBiomass}
\alias{ggPairsBiomass}
\title{ggPairsBiomass
plots pairs plot of biomass among species and provided Kendall correlation
tau}
\usage{
ggPairsBiomass(myFitCaNmod, species = NULL, logscale = TRUE)
}
\arguments{
\item{myFitCaNmod}{result sent by \link{fitmyCaNmod}}

\item{species}{the name (or a vector of name) of the species of interest
by default, all species}

\item{logscale}{should biomass be log10 transformed (default yes)}
}
\value{
a ggplot
}
\description{
ggPairsBiomass
plots pairs plot of biomass among species and provided Kendall correlation
tau
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)
ggPairsBiomass(res)

}
