% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSatiatInertia.R
\name{ggSatiatInertia}
\alias{ggSatiatInertia}
\title{ggSatiatInertia
display whether species were closed to satiation or inertia bounds}
\usage{
ggSatiatInertia(mysampleCaNmod, species = NULL, years = NULL, frac = 1)
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{species}{the species to plot (if null, default, all species)}

\item{years}{years to be plotted (default all)}

\item{frac}{fraction of points to be plot (default all)}
}
\value{
a ggplot
}
\description{
ggSatiatInertia
display whether species were closed to satiation or inertia bounds
}
\details{
for each species, a point stands for a time step and an iteration of the
MCMC. The x axis stands for the inertia constraint: a value close to 1
indicates that the growth (resp -1) was constrained by the max inertia (resp
min) bound. A value of 0 indicates no growth. Y-axis stands for satiation,
a value of 0 indicates that the species has not eaten while a value of 1
indicates that the species was close to satiation.
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaN(myCaNmod, 100)
ggSatiatInertia(res)
}
