% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSatiation.R
\name{ggSatiation}
\alias{ggSatiation}
\title{ggSatiation
plots incoming trophic fluxes (consumption) versus biomass, alongside the
satiation constraints (red dashed line)}
\usage{
ggSatiation(mysampleCaNmod, species = NULL, years = NULL, frac = 1)
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{species}{the name (or a vector of name) of the species of interest
by default, all species}

\item{years}{years to be plotted (default all)}

\item{frac}{fraction of points to be plot (default all)}
}
\value{
a ggplot
}
\description{
ggSatiation
plots incoming trophic fluxes (consumption) versus biomass, alongside the
satiation constraints (red dashed line)
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaN(myCaNmod, 100)
ggSatiation(res)

}
