% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTopDownBottomUp.R
\name{ggTopDownBottomUp}
\alias{ggTopDownBottomUp}
\title{ggTopDownBottomUp
display a diagram to detect bottom up control}
\usage{
ggTopDownBottomUp(mysampleCaNmod, species = NULL, years = NULL, frac = 1)
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{species}{a named list with for each species of interest (names), a
vector of predators to consider (if null, default, all species and all
predators)}

\item{years}{years to be plotted (default all)}

\item{frac}{fraction of points to be plot (default all)}
}
\value{
a ggplot
}
\description{
ggTopDownBottomUp
display a diagram to detect bottom up control
}
\details{
the idea of this plot is to show whether a species is controlled by its
predators
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaN(myCaNmod, 100)
ggTopDownBottomUp(res)
}
