% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggViolin.R
\name{ggViolin}
\alias{ggViolin}
\title{ggViolin}
\usage{
ggViolin(myFitCaNmod, param, year, logscale = TRUE,
  xlab = "Component/Flux", ylab = "Biomass")
}
\arguments{
\item{myFitCaNmod}{result sent by \link{fitmyCaNmod}}

\item{param}{the name (or a vector of name) of a parameter (either a flow or
a biomass)}

\item{year}{the year}

\item{logscale}{flag to indicate to use a log scale (default TRUE)}

\item{xlab}{default Component Flux}

\item{ylab}{default Biomass}
}
\value{
a ggplot
}
\description{
provide a violin plot for specific years
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)
#with one series#'
#with 2 series
ggViolin(res,c("F01","HerbZooplankton"), year=1988, TRUE)

}
