% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.VecBasic.R
\name{mean.VecBasic}
\alias{mean.VecBasic}
\title{mean.VecBasic
this function is for internal use}
\usage{
\method{mean}{VecBasic}(x)
}
\arguments{
\item{x}{a VecBasic}
}
\value{
the symbolic expression corresponding to the mean of x
}
\description{
mean.VecBasic
this function is for internal use
}
