% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfCaNmod.R
\name{pdfCaNmod}
\alias{pdfCaNmod}
\title{pdfCaNmod
creates a pdf summary of the case study}
\usage{
pdfCaNmod(myCaNmod, output_file = "summaryCaNmod.pdf")
}
\arguments{
\item{myCaNmod}{a CaNmod oject}

\item{output_file}{name of the output file}
}
\description{
pdfCaNmod
creates a pdf summary of the case study
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
pdfCaNmod(myCaNmod)
}
