% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolytope2D.R
\name{plotPolytope2D}
\alias{plotPolytope2D}
\title{plotPolytope2D
polt the possible values of a multimensional convex polytope by A.x<=b and
C.x=v or by a CaNmod object in 2D}
\usage{
plotPolytope2D(x, params = c(1, 2), progressBar = TRUE)
}
\arguments{
\item{x}{either a CaNmod oject or a named list with at least a matrix A and
a vector b (A.x<=b) and optionnally a matrix C and a vector v (C.x=v)}

\item{params}{a vector of length 2 corresponding to the index of the
parameters (default, two first)}

\item{progressBar}{should a progress bar be displayed (default TRUE)}
}
\value{
a ggplot
}
\description{
plotPolytope2D
polt the possible values of a multimensional convex polytope by A.x<=b and
C.x=v or by a CaNmod object in 2D
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- plotPolytope2D(list(A = A, b = b))

}
