% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolytope2DCaNmod.R
\name{plotPolytope2DCaNmod}
\alias{plotPolytope2DCaNmod}
\title{plotPolytope2DCaNmod
polt the possible values of the polytope 2 dimensions}
\usage{
plotPolytope2DCaNmod(myCaNmod, params)
}
\arguments{
\item{myCaNmod}{a CaNmod oject}

\item{params}{the names of the parameters}
}
\value{
a ggplot
}
\description{
plotPolytope2DCaNmod
polt the possible values of the polytope 2 dimensions
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
plotPolytope2DCaNmod(myCaNmod, c("OmniZooplankton[1988]", "F01[1988]"))

}
