% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presolveLPMod.R
\name{presolveLPMod}
\alias{presolveLPMod}
\title{presolveLPMod
simplify the lp problem using lpsolveAPI}
\usage{
presolveLPMod(A, b, C = NULL, v = NULL, lower = NULL)
}
\arguments{
\item{A}{a matrix of inequality constraints A x <=b}

\item{b}{a vector A x <=b}

\item{C}{a matrix of equality constraints C x = v, should be null in the
absence of such constraints}

\item{v}{a vector of equality constraints C x = v, should be null in the
absence of such constraints}

\item{lower}{minimal bounds for paramaters, by default set to zero}

\item{fixed}{value of potentially fixed values (Null if no parameter are
fixed)}
}
\value{
a vector corresponding to the centroid of the polytope
}
\description{
presolveLPMod
simplify the lp problem using lpsolveAPI
}
