% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatX.R
\name{reformatX}
\alias{reformatX}
\title{reformatX
check that x parameter is in correct format and returns it as a list}
\usage{
reformatX(x)
}
\arguments{
\item{x}{either a CaNmod or a named list}
}
\value{
a list with A, b, C and v
}
\description{
reformatX
check that x parameter is in correct format and returns it as a list
}
