% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_solve.R
\name{ROI_solve}
\alias{ROI_solve}
\title{ROI_solve
calls \code{\link[ROI]{ROI_solve}} from ROI except for lpsolve}
\usage{
ROI_solve(x, solver, control = list())
}
\arguments{
\item{x}{an OP object from ROI}

\item{solver}{the solver to use}

\item{control}{a list of control parameter}
}
\value{
same as \code{\link[ROI]{ROI_solve}}
}
\description{
ROI_solve
calls \code{\link[ROI]{ROI_solve}} from ROI except for lpsolve
}
