% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCaN.R
\name{sampleCaN}
\alias{sampleCaN}
\title{sampleCaN}
\usage{
sampleCaN(
  myCaNmod,
  N,
  nchain = 1,
  ncore = 1,
  thin = 1,
  method = "gibbs",
  lastF = FALSE
)
}
\arguments{
\item{myCaNmod}{a CaNmod or a sampleCaNmod. In this latter case, adapative
phase and discarding phases won't be necessary}

\item{N}{the number of samples required}

\item{nchain}{the number of mcmc chains}

\item{ncore}{number of cores to use}

\item{thin}{thinning interval}

\item{method}{one of gibbs (default) or hitandrun}

\item{lastF}{should flow for last year be simulated (default = FALSE)}
}
\value{
a sampleCaNmod object which contains three elements
\itemize{
 \item{"CaNmod"}{the CaNmod object descring the model}
 \item{"mcmc"}{\code{\link[coda]{mcmc.list}}}
 \item{"covMat"}{the estimate of the covariance matrix in the first chain
 which can be used to samples new iterations without adaptation nor
 discarding phase}
}
}
\description{
sample the polytope corresponding to the CaNmod model
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaN(myCaNmod, 100)

}
