% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_constraint.R
\name{toggle_constraint}
\alias{toggle_constraint}
\title{toggle_constraint}
\usage{
toggle_constraint(myCaNmod, constr)
}
\arguments{
\item{myCaNmod}{a CaNmod object with following elements}

\item{constr}{a vector of strings identifying the constraints to be toogled.
The elements will be compared to the beginning of the line names of
matrices A and C (regular expressions).}
}
\value{
an updated CaNmod object
}
\description{
Activate or inactivate a constraint.
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
toggle_constraint(myCaNmod, "C02")

}
