% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_constraint.R
\name{treat_constraint}
\alias{treat_constraint}
\title{treat_constraint}
\usage{
treat_constraint(myconstraint, symbolic_enviro, yr = NULL,
  name_constr = NULL)
}
\arguments{
\item{myconstraint}{a string corresponding to a constraint}

\item{symbolic_enviro}{an environment that stores all symbolic computation}

\item{yr}{the year ranges in which the constraint apply}

\item{name_constr}{a string naming the constraint}
}
\value{
a matrix that correspond to the coefficient of the constraints, first
column corresponds to the -intercept and each line to a constraint (year)
}
\description{
This is an internal function used to convert a string constraint into a
symbolic expression
}
