% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCaNmod_toXLS.R
\name{writeCaNmod_toXLS}
\alias{writeCaNmod_toXLS}
\title{writeCaNmod_toXLS}
\usage{
writeCaNmod_toXLS(file, myCaNmod)
}
\arguments{
\item{file}{the name of the export file}

\item{myCaNmod}{a CaNmod object to be exported}
}
\description{
writes an xlsx file that would correspond to the template of the CaNMod
object
}
\examples{
myCaNmod <- buildCaN(system.file("extdata",
 "CaN_template_mini.xlsx", package = "RCaN"))
 writeCaNmod_toXLS(paste(tempdir(),
                         "/export.xlsx",
                         sep = ""),
                   myCaNmod)

}
