% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_CaN_mod.R
\name{build_CaNmod}
\alias{build_CaNmod}
\title{build_CaNmod}
\usage{
build_CaNmod(file)
}
\arguments{
\item{file}{the name of the file enclosing the model description}
}
\value{
a CaNmod object with following elements
\itemize{
 \item{"components_param"}{the table of components description}
 \item{"species"}{the name of the species}
 \item{"fluxes_def}{the table of fluxes definition}
 \item{"ntstep"}{the number of time steps}
 \item{"data_series_name"}{the names of the data series}
 \item{"constraints"}{the table of constraints description}
 \item{"H"}{the H matrix from (I-H).B+N.F}
 \item{"N"}{the N matrix from (I-H).B+N.F}
 \item{"A"}{matrix of constraints A.x<=b}
 \item{"b"}{vector of constraints A.x<=b}
 \item{"C"}{matrix of constraints C.x=v}
 \item{"v"}{vector of constraints C.x=v}
 \item{"L"}{matrix of B=L.F, since B0 is a parameter, there is no M}
 \item{"symbolic_enviro"}{an environment storing all symbolic objects
 required for the computation}
}
}
\description{
function that reads the template and return build the complete model
description, including all the underlying equations
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata",
 "CaN_template_mini.xlsx", package = "RCaN"))

}
