% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPolytopeStatus.R
\name{checkPolytopeStatus}
\alias{checkPolytopeStatus}
\title{checkPolytopeStatus
check if the polytope has solution or not}
\usage{
checkPolytopeStatus(A, b, C = NULL, v = NULL)
}
\arguments{
\item{A}{the matrix of inequality A.x<=b}

\item{b}{the vector A.x<=b}

\item{C}{the matrix of equality C.x=v (default NULL for no equality)}

\item{v}{the vector of equality C.x=v (default NULL for no equality}
}
\value{
print a message to tell if the polygon is ok or not
}
\description{
checkPolytopeStatus
check if the polytope has solution or not
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- checkPolytopeStatus(A,b)

#this one is empty
C <- matrix(c(1,rep(0,n-1)),1)
v <- 3
X0 <- checkPolytopeStatus(A,b,C,v)

}
