% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmyCaNmod.R
\name{fitmyCaNmod}
\alias{fitmyCaNmod}
\title{fitmyCaNmod}
\usage{
fitmyCaNmod(myCaNmod, N, nchain = 1, ncore = 1)
}
\arguments{
\item{myCaNmod}{a CaNmod object with following elements}

\item{N}{the number of samples required}

\item{nchain}{the number of mcmc chains}

\item{ncore}{number of cores to use}
}
\value{
a \code{\link[coda]{mcmc.list}}
}
\description{
fit the CaNmod model
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx", package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)

}
