% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmyCaNmod.R
\name{fitmyCaNmod}
\alias{fitmyCaNmod}
\title{fitmyCaNmod}
\usage{
fitmyCaNmod(myCaNmod, N, nchain = 1, ncore = 1, thin = 1,
  method = "gibbs", lastF = FALSE)
}
\arguments{
\item{myCaNmod}{a CaNmod object with following elements}

\item{N}{the number of samples required}

\item{nchain}{the number of mcmc chains}

\item{ncore}{number of cores to use}

\item{thin}{thinning interval}

\item{method}{one of gibbs (default) or hitandrun}

\item{lastF}{should flow for last year be simulated (default = FALSE)}
}
\value{
a fitCaNmod object which contains two elements
\itemize{
 \item{"CaNmod"}{the CaNmod object descring the model}
 \item{"mcmc"}{\code{\link[coda]{mcmc.list}}}
}
}
\description{
fit the CaNmod model
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)

}
