% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCaNmod.R
\name{ggCaNmod}
\alias{ggCaNmod}
\title{ggCaNmod
produces a ggraph representing the flow foodweb}
\usage{
ggCaNmod(myCaNmod)
}
\arguments{
\item{myCaNmod}{a CaNmod oject}

\item{...}{Additional arguments}
}
\value{
a ggraph
}
\description{
ggCaNmod
produces a ggraph representing the flow foodweb
}
\examples{
require(ggraph)
require(ggplot2)
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx", package = "RCaN"))
ggCaNmod(myCaNmod)
}
