% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDiet.R
\name{ggDiet}
\alias{ggDiet}
\title{ggDiet}
\usage{
ggDiet(myFitCaNmod, species, barplot = TRUE)
}
\arguments{
\item{myFitCaNmod}{result sent by \link{fitmyCaNmod}}

\item{species}{the name (or a vector of name) of the species of interest}

\item{barplot}{if TRUE a barplot (default), else a violin plot}
}
\value{
a ggplot
}
\description{
provide a ggplot of diet, barplot provides an average diet while violin plot
provides a distribution over iterations and years
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)
ggDiet(res,"OmniZooplankton")

}
