% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGrowth.R
\name{ggGrowth}
\alias{ggGrowth}
\title{ggGrowth
plots biomass growth as a function of biomass
provides a distribution over iterations and years}
\usage{
ggGrowth(mysampleCaNmod, species = NULL, years = NULL)
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{species}{the name (or a vector of name) of the species of interest
by default, all species}

\item{years}{years to be plotted (default all)}
}
\value{
a ggplot
}
\description{
ggGrowth
plots biomass growth as a function of biomass
provides a distribution over iterations and years
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaN(myCaNmod, 100)
ggGrowth(res)

}
