% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggResult.R
\name{ggResult}
\alias{ggResult}
\title{ggResult}
\usage{
ggResult(mysampleCaNmod, param, plot_series = TRUE,
  ylab = "Biomass/Flux")
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{param}{the name (or a vector of name) of a parameter (either a flow or
a biomass)}

\item{plot_series}{if yes, draw three example of trajectories (default TRUE)}

\item{ylab}{default Biomass Flux}
}
\value{
a ggplot
}
\description{
provide a ggplot of the fit
}
\details{
the line corresponds to median of the mcmc simulation, the ribbon
corresponds to quantiles 2.5% and 97.5%
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaNmod(myCaNmod, 100)
#with one series
ggResult(res,"F01", TRUE)

#with 2 series
ggResult(res,c("F01","HerbZooplankton"), TRUE)

}
