% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSatiation.R
\name{ggSatiation}
\alias{ggSatiation}
\title{ggSatiation
plots incoming trophic fluxes versus biomass}
\usage{
ggSatiation(myFitCaNmod, species = NULL)
}
\arguments{
\item{myFitCaNmod}{result sent by \link{fitmyCaNmod}}

\item{species}{the name (or a vector of name) of the species of interest
by default, all species}
}
\value{
a ggplot
}
\description{
ggSatiation
plots incoming trophic fluxes versus biomass
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)
ggSatiation(res)

}
