% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTrophicRelation.R
\name{ggTrophicRelation}
\alias{ggTrophicRelation}
\title{ggTrophicRelation}
\usage{
ggTrophicRelation(myFitCaNmod, species = NULL)
}
\arguments{
\item{myFitCaNmod}{result sent by \link{fitmyCaNmod}}

\item{species}{the name (or a vector of name) of the species of interest
by default, all species}
}
\value{
a ggplot
}
\description{
plots fluxes from prey to predators
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- build_CaNmod(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- fitmyCaNmod(myCaNmod, 100)
ggTrophicRelation(res)

}
