% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggViolin.R
\name{ggViolin}
\alias{ggViolin}
\title{ggViolin
provides a violin plot of the distribution of flux or biomass
for specific years}
\usage{
ggViolin(mysampleCaNmod, param, year, logscale = TRUE,
  xlab = "Component/Flux", ylab = "Distribution")
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{param}{the name (or a vector of name) of a parameter (either a flux or
a biomass)}

\item{year}{the year}

\item{logscale}{flag to indicate to use a log scale (default TRUE)}

\item{xlab}{default Component Flux}

\item{ylab}{default Distribution}
}
\value{
a ggplot
}
\description{
ggViolin
provides a violin plot of the distribution of flux or biomass
for specific years
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaN"))
res <- sampleCaN(myCaNmod, 100)
#with 2 series
ggViolin(res,c("F01","HerbZooplankton"), year=1988, TRUE)

}
