% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolytope2D.R
\name{plotPolytope2D}
\alias{plotPolytope2D}
\title{plotPolytope2D
polt the possible values of a multimensional convex polytope by A.x<=b and
C.x=v for 2 dimensions}
\usage{
plotPolytope2D(A, b, C = NULL, v = NULL, params = c(1, 2))
}
\arguments{
\item{A}{the matrix of inequality A.x<=b}

\item{b}{the vector A.x<=b}

\item{C}{the matrix of equality C.x=v (default NULL for no equality)}

\item{v}{the vector of equality C.x=v (default NULL for no equality}

\item{params}{a vector of length 2 corresponding to the index of the
parameters}
}
\value{
a ggplot
}
\description{
plotPolytope2D
polt the possible values of a multimensional convex polytope by A.x<=b and
C.x=v for 2 dimensions
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- plotPolytope2D(A,b)

}
